add option -f (force ln)

+++ setup.sh
@@ -7,13 +7,20 @@ TMUX_CMDS=( tmuxx tmux-pbcopy )
 ZSH_CUSTOM_DIR=$HOME/.oh-my-zsh/custom
 CMD_DIR=/usr/local/bin
 
+FORCE_OPTION="-f"
+
 # HOME
 for file in ${DOT_FILES[@]}
 do
-	if [ -a $HOME/$file ]; then
-		echo "既にファイルが存在します: $file"
+	if [ $FORCE_OPTION != "$1" ]; then
+		if [ -a $HOME/$file ]; then
+			echo "既にファイルが存在します: $file"
+		else
+			ln -s $HOME/dotfiles/$file $HOME/$file
+			echo "シンボリックリンクを貼りました: $file"
+		fi
 	else
-		ln -s $HOME/dotfiles/$file $HOME/$file
+		ln -fs $HOME/dotfiles/$file $HOME/$file
 		echo "シンボリックリンクを貼りました: $file"
 	fi
 done
@@ -21,10 +28,15 @@ done
 # oh-my-zsh custom
 for file in ${CUSTOM_ZSH_FILES[@]}
 do
-	if [ -a $ZSH_CUSTOM_DIR/$file ]; then
-		echo "既にファイルが存在します: $file"
+	if [ $FORCE_OPTION != "$1" ]; then
+		if [ -a $ZSH_CUSTOM_DIR/$file ]; then
+			echo "既にファイルが存在します: $file"
+		else
+			ln -s $HOME/dotfiles/$file $ZSH_CUSTOM_DIR/$file
+			echo "シンボリックリンクを貼りました: $file"
+		fi
 	else
-		ln -s $HOME/dotfiles/$file $ZSH_CUSTOM_DIR/$file
+		ln -fs $HOME/dotfiles/$file $ZSH_CUSTOM_DIR/$file
 		echo "シンボリックリンクを貼りました: $file"
 	fi
 done
@@ -32,10 +44,15 @@ done
 # tmux
 for file in ${TMUX_CMDS[@]}
 do
-	if [ -a $CMD_DIR/$file ]; then
-		echo "既にファイルが存在します: $file"
+	if [ $FORCE_OPTION != "$1" ]; then
+		if [ -a $CMD_DIR/$file ]; then
+			echo "既にファイルが存在します: $file"
+		else
+			ln -s $HOME/dotfiles/$file $CMD_DIR/$file
+			echo "シンボリックリンクを貼りました: $file"
+		fi
 	else
-		ln -s $HOME/dotfiles/$file $CMD_DIR/$file
+		ln -fs $HOME/dotfiles/$file $CMD_DIR/$file
 		echo "シンボリックリンクを貼りました: $file"
 	fi
 done
